Reported Frag scroll Location data sent to ReportFragment map button change manifest changed, orientation=portrait

+++ app/src/main/java/com/example/reinstall_app/app_data/ReportedProblem.java
@@ -9,7 +9,7 @@ public class ReportedProblem {
 
     String description, userName, problemType, city, suburb;
     Date date;
-    int x, y;
+    double x, y;
 
 
     public Date getDate() {
@@ -44,19 +44,19 @@ public class ReportedProblem {
         this.suburb = suburb;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
